feat: enhance merge workflow with outputs and branch protection checks
With this change, if the branch is not protected by fast-forward checks,
then the job will succeed, and simply not run the merge logic. This is
a change from the current process, which will exit the workflow, marking
the workflow as failed.
Having failed workflows is undesirable, as it indicates to the developer
that somthing needs to be fixed, when in fact, the process has
functioned correctly.